Remove all lengthy CHANGELOG, TODO type comments from source files
I would humbly like to suggest the project not include extremely lengthy
CHANGELOG and TODO type comments in the code. Some of these comments are
describing changes from 2005. In my opinion, this is not the best place
to store such comments it is better served by other tools such as the
git log or a CHANGELOG file (http://keepachangelog.com/en/0.3.0/). Big
TODO items are better served through
issues (https://github.com/decalage2/olefile/issues) or a project
roadmap.

For new contributors reading the code, these comments are very
unconventional and make it difficult to read the code as there is so
much surrounding noise. I have removed them to help tidy up the code and
bring it in line with conventions and other Python libraries.