Move non-library package data outside of Python package
This avoids having the files themselves installed as Python files used
by the package. As these files are not used by the actual Python code,
no need to install alongside the library. All files remain a part of the
source package, are just not installed.

The license file is now installed using the wheel package's supported
method. Added to [metadata] to setup.cfg with the property license-file,
with this, the license is now installed in to the dist-info directory.