Ensure that self.fp is closed at destruct
If we opened it, we are responsible for closing it.

This also rids us of ResourceWarnings popping up in newer python versions.

Of course, actually, the creator of an OleFileIO is repsonsible for calling
close() before it goes out of scope. But if the error happens in open()
called from __init__, then caller has no chance to call close().